receiver/requirements.txt
aiohttp==3.8.5
python-dotenv==1.0.0
pydantic==2.1.1
SQLAlchemy==2.0.19
asyncpgsa==0.27.1
alembic==1.11.2
psycopg2-binary==2.9.7
pytest==7.4.0
pytest-asyncio==0.21.1
